requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib-metadata==6.4.1
joblib==1.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-hub==0.13.0
termcolor==1.1.0
tfp-nightly==0.20.0.dev20230419
threadpoolctl==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.12.1
zipp==3.15.0
